refactor(treeNode): simplify component
