+++ package.json
@@ -66,6 +66,7 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.31",
     "karma-webpack": "^2.0.4",
+    "lodash": "^4.17.11",
     "mocha": "^3.4.2",
     "node-libs-browser": "^2.0.0",
     "react": "^16.8.5",
