+++ src/index.js
@@ -1,5 +1,3 @@
-
-
 import Treebeard from './components/treebeard';
 import decorators from './components/decorators';
 import animations from './themes/animations';
