+++ test/src/components/treebeard-tests.js
@@ -5,7 +5,7 @@ import React from 'react';
 import TestUtils from 'react-dom/test-utils';
 
 import defaultDecorators from '../../../src/components/decorators';
-import TreeNode from '../../../src/components/node';
+import TreeNode from '../../../src/components/TreeNode';
 import Treebeard from '../../../src/components/treebeard';
 import defaultAnimations from '../../../src/themes/animations';
 import defaultTheme from '../../../src/themes/default';
